Move IndexStoreDBError to a separate file
